Remove 4 unboxed warnings
Use Integer.intValue to explicitly convert to an int

+++ org.eclipse.jgit/src/org/eclipse/jgit/revplot/PlotCommitList.java
@@ -225,8 +225,8 @@ public class PlotCommitList<L extends PlotLane> extends
 		if (blockedPositions.get(commit.lane.getPosition())) {
 			int newPos = -1;
 			for (Integer pos : freePositions)
-				if (!blockedPositions.get(pos)) {
-					newPos = pos;
+				if (!blockedPositions.get(pos.intValue())) {
+					newPos = pos.intValue();
 					break;
 				}
 			if (newPos == -1)
